extract CopyToClipboardTooltip component
