+++ apps/dashboard/volto/Tooltip/index.ts
@@ -1,3 +1,7 @@
+export {
+  CopyToClipboardTooltip,
+  useCopyToClipboardState,
+} from './CopyToClipboardTooltip';
 export {Tooltip} from './Tooltip';
 export {TooltipsManagerProvider} from './TooltipsManager';
 export {useTooltip} from './useTooltip';
